clear: Clear out the dynamic logic and rebuild the code status
